update time type

+++ backend/src/main/java/cookme/recipesmodel/RecipesDto.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public record RecipesDto(String name,
                          String description,
-                         Integer time,
+                         double time,
                          String imageUrl,
                          String preparation,
                          Status status,
